refactor 'swipe-controls' patch
